Init http model for get utxo and transfer

+++ server/internal/domain/chain_tx/entity.go
@@ -6,7 +6,7 @@ type ChainTx struct {
 	common.Entity `json:",inline"`
 	TxID          []byte `json:"tx_id"`
 	AnchorTx      []byte `json:"anchor_tx"`
-	ChainFees     int64  `json:"chain_fees,omitempty"`
+	ChainFees     int32  `json:"chain_fees,omitempty"`
 	BlockHeight   int32  `json:"block_height,omitempty"`
 	BlockHash     []byte `json:"block_hash,omitempty"`
 }
